# openapi.yaml
openapi: 3.1.0
info:
  title: physique-analytics API
  version: 0.1.0
  description: |
    トレーニング・栄養・体組成を統合して分析するアプリの API。

    このファイルが **型の唯一の正** であり、Go サーバ・TypeScript クライアント・
    Swift クライアントのコードはすべてここから生成する（ADR-0007）。
    手で型を書かないこと。

    ## 日付の扱い
    `date` はすべて **JST（Asia/Tokyo）における日付**（ADR-0013）。
    端末のタイムゾーンに依存しない。クライアントが JST で計算して送る。

    ## 同期と削除
    すべてのリソースが `updatedAt` / `deletedAt` を持つ（ADR-0014）。
    競合は `updatedAt` が新しい方を採用する（Last Write Wins）。
    削除は物理削除せず `deletedAt` を立てる。
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: https://api.example.com
    description: production
  - url: http://localhost:8080
    description: local
security:
  - bearerAuth: []
...
paths:
...
  /v1/measurements:
    get:
      operationId: listMeasurements
      tags: [body]
      summary: 周囲長の一覧
      description: |
        要件 B-02。**起床直後・食事前・同じメジャー・同じ強さ**で測る
        （docs/02-データモデル.md）。日中は食事と姿勢でウエストが2〜3cm動く。
      parameters:
        - $ref: "#/components/parameters/DateFrom"
        - $ref: "#/components/parameters/DateTo"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties:
                  items:
                    type: array
                    items: { $ref: "#/components/schemas/BodyMeasurement" }
        "401": { $ref: "#/components/responses/Unauthorized" }
    put:
      operationId: putMeasurement
      tags: [body]
      summary: 周囲長の登録・更新
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/BodyMeasurementInput" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/BodyMeasurement" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "422": { $ref: "#/components/responses/ValidationFailed" }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        Supabase Auth が発行した JWT。API 側は JWKS で検証し、
        `sub` クレームから user_id を取り出す（インフラ設計 §8）。
  parameters:
...
    DateFrom:
      name: from
      in: query
      schema: { type: string, format: date }
      description: JST の日付（ADR-0013）
    DateTo:
      name: to
      in: query
      schema: { type: string, format: date }
...
  schemas:
...
    Timestamps:
      type: object
      required: [createdAt, updatedAt]
      properties:
        createdAt: { type: string, format: date-time }
        updatedAt:
          type: string
          format: date-time
          description: 競合解決に使う（ADR-0014）
        deletedAt:
          type: [string, "null"]
          format: date-time
          description: 論理削除。物理削除しない（ADR-0014）
...
    BodyMeasurement:
      allOf:
        - $ref: "#/components/schemas/Timestamps"
        - type: object
          required: [id, date]
          properties:
            id: { type: string, format: uuid }
            date: { type: string, format: date }
            neckCm: { type: [number, "null"], exclusiveMinimum: 0, maximum: 100 }
            shoulderCm: { type: [number, "null"], exclusiveMinimum: 0, maximum: 200 }
            chestCm: { type: [number, "null"], exclusiveMinimum: 0, maximum: 200 }
            waistNavelCm: { type: [number, "null"], exclusiveMinimum: 0, maximum: 200 }
            hipCm: { type: [number, "null"], exclusiveMinimum: 0, maximum: 200 }
            armRCm: { type: [number, "null"], exclusiveMinimum: 0, maximum: 100 }
            thighRCm: { type: [number, "null"], exclusiveMinimum: 0, maximum: 150 }
            calfRCm: { type: [number, "null"], exclusiveMinimum: 0, maximum: 100 }
    BodyMeasurementInput:
      type: object
      required: [date]
      properties:
        date: { type: string, format: date }
        neckCm: { type: [number, "null"], exclusiveMinimum: 0, maximum: 100 }
        shoulderCm: { type: [number, "null"], exclusiveMinimum: 0, maximum: 200 }
        chestCm: { type: [number, "null"], exclusiveMinimum: 0, maximum: 200 }
        waistNavelCm: { type: [number, "null"], exclusiveMinimum: 0, maximum: 200 }
        hipCm: { type: [number, "null"], exclusiveMinimum: 0, maximum: 200 }
        armRCm: { type: [number, "null"], exclusiveMinimum: 0, maximum: 100 }
        thighRCm: { type: [number, "null"], exclusiveMinimum: 0, maximum: 150 }
        calfRCm: { type: [number, "null"], exclusiveMinimum: 0, maximum: 100 }
        updatedAt: { type: string, format: date-time }
...
    Problem:
      type: object
      description: RFC 7807 Problem Details
      required: [type, title, status]
      properties:
        type: { type: string, format: uri, default: "about:blank" }
        title: { type: string }
        status: { type: integer }
        detail: { type: string }
        instance: { type: string }
        errors:
          type: array
          description: バリデーションエラーの詳細
          items:
            type: object
            required: [field, message]
            properties:
              field: { type: string }
              message: { type: string }
  responses:
    BadRequest:
      description: リクエストが不正
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }
    Unauthorized:
      description: 認証されていない
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }
...
    ValidationFailed:
      description: |
        入力値の検証に失敗（要件 §4.8）。
        体重30〜200kg、未来日の禁止、RIR 0〜10 など。
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }
...